IGAPP-852: use spec reporter & add cmd option to select specific browser
