+++ docs/e2e-tests.md
@@ -16,7 +16,6 @@ Browserstack provides you with a large set of available operating systems, brows
 To see and run the E2E tests on Browserstack you need a Browserstack account, with access to our organization.
 You will need to set `E2E_BROWSERSTACK_KEY` and `E2E_BROWSERSTACK_USER` as environment variables, which you can find when logging in to Browserstack and selecting either the Automate or AppAutomate tab.
 If you don't have this access you can still execute the tests locally with `yarn test:web` or `yarn test:native`.
-For the browserstack tests in web there are several configurations to run the tests in the specified browsers. Preferably you should run these from the commandline, e.g. `yarn wdio run web/browserstack/wdio-browserstack-safari.conf.ts` from the project root. There is also a configuration in the `package.json`, which runs all of the browserstack configurations sequentially.
 For native browserstack tests you also have to set the environment variable `E2E_CONFIG` to either `android` or `ios`.
 You also have to set the `E2E_BROWSERSTACK_APP` variable, which corresponds to the id of a `*.apk` or `*.ipa`, which is already uploaded to browserstack.
 To get such id, you either upload an `*.apk` or `*.ipa` to browserstack or use an already existing one (Checkout other jobs. Using the same `*.apk` will not influence the other test). The id looks something like this: `bs://8ad23ac842b0ed12dafff0461bc3a0b98484234f`.
@@ -47,7 +46,15 @@ yarn workspace e2e test:web
 > Run the following command in your terminal:
 > `/usr/bin/safaridriver --enable`
 
-This will launch an automated chrome session where you can follow the test execution. If your don't have chrome installed you can adjust the `browserName` (firefox, edge, safari) in the `wdio.conf.ts` in the E2E-test web subdirectory. Chromium does not seem to work currently. Safari also does not work on the current version (14.1).
+This will launch an automated chrome session where you can follow the test execution. If your don't have chrome
+installed you can append `--firefox` or any other browser of your choice to the command. Chromium does not seem to work
+currently. Safari also does not work on the current version (14.1).
+
+To run a single spec file you can run
+
+```
+yarn workspace e2e test:web --spec regexp
+```
 
 ### Native
