+++ e2e-tests/package.json
@@ -18,7 +18,7 @@
     "@wdio/browserstack-service": "^7.16.6",
     "@wdio/cli": "^7.16.6",
     "@wdio/jasmine-framework": "^7.16.6",
-    "@wdio/junit-reporter": "^7.16.6",
+    "@wdio/spec-reporter": "^7.16.9",
     "@wdio/local-runner": "^7.16.6",
     "@wdio/selenium-standalone-service": "^7.16.6",
     "@wdio/types": "^7.16.3",
