--- e2e-tests/web/browserstack/wdio-browserstack-firefox.conf.ts
@@ -1,5 +1,0 @@
-import { config as defaultConfig } from './wdio-browserstack.conf'
-
-const config = { ...defaultConfig }
-config.capabilities = config.capabilities.filter(cap => cap.browserName?.toLowerCase() === 'firefox')
-export { config }
