--- e2e-tests/web/browserstack/wdio-browserstack.conf.ts
@@ -1,14 +1,0 @@
-import { browserstackCapabilities } from '../capabilities'
-import { config as defaultConfig } from '../wdio.conf'
-
-export const config = Object.assign(defaultConfig, {
-  maxInstances: 1,
-
-  user: process.env.E2E_BROWSERSTACK_USER,
-  key: process.env.E2E_BROWSERSTACK_KEY,
-
-  capabilities: Object.values(browserstackCapabilities),
-
-  services: [['browserstack', { browserstackLocal: true }]],
-  host: 'hub.browserstack.com'
-})
